Construct `BenchmarkRunner` within `benchmark_replication`; move `evaluate_oracle` from runner to problem (#3034)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3034

Context:

Moving the runner off of the `BenchmarkProblem` and instead constructing it within `BenchmarkReplication`: In addition to making it simpler for the user to specify the problem and making the code flatter and easier to understand, this will enable the runner to depend on both the `BenchmarkProblem` and `BenchmarkMethod`, so that it can take advantage of the parallelism stored on the runner.

Moving `evaluate_oracle` from the runner to the problem is the right thing to do in principle, since ideally the runner would only need to be used within the BayesOpt loop. Also, there could conceivably be a world where the oracle value is the "best" value (e.g. the lowest point on a learning curve), and that woulud need to depend on the `OptimizationConfig`, which the `BenchmarkProblem` wouldn't need to know about.

This diff:
* Removes `runner` from `BenchmarkProblem`
* Starts constructing the runner in `benchmark_replication`
* Moves `evaluate_oracle` from runner to problem; moves `target_fidelity_and_task` from runner to problem in order to make that possible

Reviewed By: saitcakmak

Differential Revision: D65503068