Set Scheduler logging level appropriately (#3028)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3028

Currently, setting the level of DEBUG or lower via the `logging_level` in `SchedulerOptions` isn't working as expected  because that level is never passed as an argument when retrieving the logge.  The default value of INFO is used instead and because of this , DEBUG logs were being filtered out

Also, it looks like we are currently setting **root stream handler's level** with `logging_level` from `SchedulerOptions` - I don't think this should be the case since the root stream handler is a global handler that handles log messages from all loggers in the application. The correct intention is to set the logging level of the `Scheduler` logger specifically.

Reviewed By: bernardbeckerman

Differential Revision: D65555729

fbshipit-source-id: 30fe48f8b030eaba3dcea93453a6bdbb5047b1b3