Construct `BenchmarkRunner` automatically instead of requring it to be passed to `BenchmarkProblem` (#3022)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3022

**Context**:

This is an intermediate step. In this diff, we no longer require a user constructing a `BenchmarkProblem` to construct a `BenchmarkRunner`, instead asking the user for the arguments to `BenchmarkRunner`, `test_function` and `noise_std`. (It also requires a `search_space_digest`, which can be computed from the `search_space` on the problem.) In the next diff, the `BenchmarkRunner` will move off the `BenchmarkProblem` entirely. That will eventually allow the runner to depend on attributes that live on `BenchmarkMethod`, such as parallelism, in turn unblocking asynchronous benchmarking.

**This diff**:
* Adds arguments `test_function` and `noise_std` to `BenchmarkProblem`
* Removes argument `runner` from `BenchmarkProblem`
* Automatically constructs the runner in the `BenchmarkProblem`'s post-init
* Updates call sites

Reviewed By: Balandat

Differential Revision: D65498908

fbshipit-source-id: bc87a8b8b068ba6c092ed73f4a9d5ab6df135ba0