Expand range parameters in model fitting space to cover training data (#3040)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3040

Separates concepts of "modeling space" and "search space". Modeling space is stored where it was before, in "_model_space". Search space is stored in a new attribute, "_search_space". (Storage needs to be updated for that, TBD).

By default, modeling space will be expanded to cover the full range of RangeParameter values seen in the training data. This means that training data will never be marked OOD for falling outside RangeParameter box bounds.

The updated bounds will be passed along to Botorch, which means that priors on lengthscales will apply to the expanded modeling space, not the optimization search space.

At gen time, we will still respect the search space bounds as given.

Reviewed By: saitcakmak

Differential Revision: D64879752

fbshipit-source-id: 1a5de3728f47fcc5ad6dba6178815b8f8eed6412