Move Models registry backwards compatibility support into storage layer (#3043)
Summary:

The deprecated models have been raising a `DeprecationWarning` for about a year, so loss of the ability to call `Models.FULLYBAYESIAN` is not the backwards compatibility we want to support here. We're rather interested in being able to load old experiments from the DB. Having the BC support in storage layer enables this and eliminates the need for lots of class properties (deprecated in Py3.13) on `Models`.

Reviewed By: Balandat

Differential Revision: D65608756