Tutorials: Delete previously converted tutorials
I noticed while testing that the filesize of my converted tutorials directory blew up to over 1GB. This was happening because the plot data is stored using unique file names each time, resulting in the previous plot data never getting overwritten. Instead of relying on overwriting the previous files let's just wipe everything at the start.

I specifically only wipe subdirectories within the tutorials directory because there is a `index.mdx` file at the top level that is not autogenerated and we don't want to delete that.