Move oracle-related logic from `BenchmarkProblem` to `benchmark.py` (#3049)
Summary:

This moves methods on `BenchmarkProblem` to functions in `benchmark.py` without modifying them beyond changing "self" to "problem." This is not that interesting in itself; I'm doing this to make D65662132 easier to review. Doing this enables constructing a runner within the "oracle" functions (without creating circular import issues), and then using functionality from `BenchmarkRunner` and `BenchmarkMetric` to generate data.

Reviewed By: saitcakmak

Differential Revision: D65679443