Simplify oracle-related logic (#3051)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3051

**Context:**

Currently, `get_oracle_experiment_from_params` clumsily constructs the dataframe that defines the trial's data, just like `BenchmarkMetric` does. This diff changes that function to use `BenchmarkRunner` and `BenchmarkMetric` to attach metadata to a trial and get metric values in the same way it is done elsewhere in benchmarking (within the optimization loop), so that the logic isn't duplicated.

This change will be especially useful when we have `MapData` (D64198634), because then it will become extra awkward to have data-processing logic in multiple places.

**This diff:**
* Removes `BenchmarkProblem.evaluate_oracle, and
Modifies `get_oracle_experiment_from_params` to
* use a `BenchmarkRunner` to generate noiseless values
* use `fetch_data` (and implicitly, `BenchmarkMetric`) to process this into metric values
* Explicitly compute oracle values by evaluating the test function at the target task and fidelity instead of doing that within `BenchmarkProblem.evaluate_oracle`

Reviewed By: saitcakmak

Differential Revision: D65662132

fbshipit-source-id: a016c73a1b9536845407f81ca103084c20b7aa76