Add ModelBridgeMethodNotImplementedError (#3068)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3068

`ModelBridge._fit_if_implemented` catches `NotImplementedError` to support subclasses that do not need to fit any models. However, `NotImplementedError` is too generic of an error type to catch and may mask other issues lower in the stack. Defining a specific class for the intentionally not implemented methods will allow us to catch only what we intend to and raise the unexpected errors.

Reviewed By: esantorella

Differential Revision: D65948196

fbshipit-source-id: 1cbce3e4610521788374c7581d269bd29fb71db4