Remove unused `_init_state` and `reset()` from `BackendSimulator` (#3075)
Summary:

Context: The `BackendSimulator` can be reset to its initial state. This functionality is not used. I would like to be able to serialize and deserialize `BackendSimulator`s in order to permit analysis of past simulations. The presence of `_init_state` makes this much harder, since there are then two states that need to be considered, the initial state and the current state.

This PR:
* Removes the attribute `_init_state` and the method `reset` from `BackendSimulator`

Differential Revision: D66100116