Ax API entry point class (#3054)
Summary:

This replaces AxClient, offering a much more streamlined and opinionated vocabulary for conducting experiments as well as a unification of our ask-tell and Metric/Runner style experimentation APIs.

The methods provided here provide are complete in that using just these an Ax user is able to successfully utilize all Ax 1.0 launch features without having to access any implementation details, ax.core classes, etc directly.

Note: There are a few methods prefixed with "set_" that we will not consider API (because we do not want to expose their arguments to the API) but will leave as "breadcrumbs" for contributors or those otherwise using Ax as a research platform and interested in extending the library.

Differential Revision: D63836188