[Docusaurus] Add archived section to versions dropdown
With an entry to redirect users to the archived version of the website. Currently points to `ax.dev` which currently serves what will be the archived version.

We will need to host the archived docs at a new page once the new version is hosted at `ax.dev`

Maybe `old.ax.dev` or `archive.ax.dev`?