Allow pending trials to complete when failure rate is exceeded (#3070)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3070

This commit allows any pending trials to be completed when the maximum permissible failure rate is exceeded. This is achieved adding a private Boolean attribute `_failure_rate_has_been_exceeded`, which causes `_get_max_pending_trials` to return returning `0` when it is true, leading the scheduler to stop adding new trials.

Reviewed By: bernardbeckerman

Differential Revision: D65943818

fbshipit-source-id: ed60f06ead08915814a2639e40e51f6908f8e0c5