Prepare for "Fix type-safety of `torch.nn.Module` instances": fbcode/a* (#3088)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3088

In D52890934 we are tightening the type signature on attributes on nn Module.  To do this, we need to
first apply Pyre fixme to all sites that were relying on the old Any.  This diff shards out all changes
to files that begin with a.

bypass-github-export-checks "codemod only skipping PR review"

Reviewed By: r-barnes

Differential Revision: D66233582

fbshipit-source-id: 7a1e85dacfd7cad64d38d4fb4234323089884d45