ContourPlot (for use in InteractionPlot) (#3095)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3095

Reimplement the Contour plot using the new Analysis setup. This is not necessarily urgent in and of itself, but it will be used in the modernized interaction plot (to promote code reuse).

Note that we have changed a number of things here which make the contour much better for the interaction analysis, and debatably make things better a la carte as well:
* **Removed second contour of SEM**: This is unnecessary for the interaction analysis and has been confusing to some of our users. Let's reimplement this only if we see demand
* **Removed markers on (x, y) location of arms**: This cleans up the plot visually quite a bit  (especially important when this is small in the interaction analysis) and overall should reduce confusion.

Overall users look at these to help understand a loose narrative about the relationship between the input and output variables and to get a general "vibe" of what the model is doing . **I think we should work to preserve visual clarity and reduce noise, even if it may come at the cost of some specificity.** If a user really wants to inspect the model precisely they should be using tools other than this contour plot.

Reviewed By: danielcohenlive

Differential Revision: D65233907

fbshipit-source-id: 890ec58ad49425b283fcea26ca166a1fe990a2a0