Rewrite flaky test_compute for Interaction plot (#3116)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3116

Unfortunately we cannot count the number of subplots present by taking the length of fig.data because data holds the *TRACES* not the subplots themselves, and each subplot may hold one or more traces. This was causing the test to be flaky.

Instead now we are just checking that at least one trace came from a bar plot and that the other plots were either Scatter (slice) or Contour

Reviewed By: paschai

Differential Revision: D66472847

fbshipit-source-id: e3cedac1ad7306ccaf12dfa1c7469bb52a6366a6