Remove temporary `first_step_is_instant` logic from BenchmarkRunner (#3124)
Summary:

This PR removes temporary logic that was added for backward compatibility but not desirable. Previously,  when using `MapData`, data from the first step in a progression would be available as soon as the trial started running, even if the runtime for that step was high. This was inconsistent with the behavior when not using `MapData`, in which case data from the first and only step would not become available instantly. This PR makes these behave consistently by making data from the first step not land until that step's runtime has elapsed.

See also the top of D66313060

Reviewed By: ltiao

Differential Revision: D66474870