Make deprecated SurrogateSpec arguments InitVars
Summary:
We do not want to keep the deprecated arguments as fields of the `SurrogateSpec`. We only convert them to `ModelConfig` in `__post_init__` and ignore them after. `InitVar` is the perfect tool for doing this.

Keeping these as fields leads to issues in other diffs where the values of the deprecated attribute goes out of sync with the specified `ModelConfig`.

Differential Revision: D66553624