Change default mapdata value in data_and_evaluations_from_raw_data (#3145)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3145

V3: Made the change directly in MapData, disregard previous diff summary. See diff comments for reasoning

Infer map_key_infos with default value NaN from the shape of the data in data_and_evaluations_from_raw_data. This method  is used in trial.update_trial_data which in turn is used in AxClient and the preview API.

Ultimately we want to make this change to prevent 0 from sneaking into MapData when a user calls Client.attach_data without a progression -- we would prefer a NaN to be there signifying no data is present.

Reviewed By: saitcakmak

Differential Revision: D66765038

fbshipit-source-id: fd8c54d49e2bc3af66f29ca27eea34d0f973348b