Implement attach_data and complete_trial (#3121)
Summary:

Implement methods needed for adding partial or complete data to a trial.

Users will call attach_data if they have some partial data (either incomplete timeseries data OR some metrics ready but others not ready) available to save onto the experiment. This will always get saved as MapData. If `progression` is provided it will be saved as the "step" value for all metrics in raw_data. If no progression term is provided then the highest value of "step" present on that trial's existing data will be used (or 0 if none is present).

Users call complete_trial if no more data will become available for the given trial. They may provide raw_data to this function (which has behavior analogous to AxClient.complete_trial) or they may not -- if they do we first call attach_data under the hood. If after the data attaching all metrics present on the optimization config have some value present then the trial is marked COMPLETED. If any are missing the trial is marked FAILED.

Differential Revision: D66512777