Implement attach_trial and attach_baseline (#3138)
Summary:

Implement a method for attaching a single arm trial and a special helper method to attach a single arm trial, name it "baseline", and attach its arm as the status quo

RFC: Is this behavior actually what we want attach_baseline to do? Or does it make more sense just to set the status quo arm and not attach anything? IMO it makes sense to attach the trial in the single-arm trial setting, but in the batch trial setting we want to attach this arm to every trial, no? That being said Im not sure we want to bifurcate behavior like this. Curious to hear what you all think.

Reviewed By: lena-kashtelyan

Differential Revision: D66664475