Replace use of MaxTrials with MinTrials in core GS
Summary:
In a previous diff in this stack, I did some more in-depth thought and realized that the implementation of MaxTrials == MinTrials, so MaxTrials is actually MinTrials.

The reason this happened was when I began the setup of GenerationStep was min_trials_observed and num_trials. I fell into the trap of considering min_trials_observed MinTrials and num_trials as MaxTrials, however these are really both implementations of MinTrials with different thresholds.So we're going to fix that now.

I'm breaking it up into a couple of diffs so it's easy to pinpoint any issues if something goes wrong


this diff handles the replacement for core gs stack

Differential Revision: D66779364