Implement get_next_trials (#3107)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3107

As titled. Get next trials takes in a maximum number of trials to generate and optionally a dict of parameter values to be fixed and generates up to that number of trials (as parallelism is available).

If Experiment is not set or Experiment is set but Optimization is not set raise an Exception. If GenerationStrategy is not set then choose one automatically.

Once generated trials are attached to the experiment and immediately marked RUNNING

Reviewed By: lena-kashtelyan

Differential Revision: D66367189

fbshipit-source-id: abe1ebd1149cc46563c9bd2accbf74a5d57b21db