Implement configure_runner, configure_metric (#3104)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3104

configure_runner and configure_metric allow users to attach custom Runners and Metrics to their experiment.

configure_runner is fairly straightforward and just sets experiment.runner

configure_metric is more complicated: given a list of IMetrics it iterates through and tries to find a metric with the same name somewhere on the experiment. In order it checks the Objective (single, MOO, or secularized), outcome constraints, then tracking metrics. If no metric with a matching name is found then the provided metric is added as a tracking metric.

Reviewed By: lena-kashtelyan

Differential Revision: D66305614

fbshipit-source-id: a6048c2b3bf16f7d8b437d30cf67aa82644f451e