Stop clipping benchmark score trace at 100, score non-monotonic traces correctly, etc (#3158)
Summary:

**Context:**

Clipping benchmark scores at 100 means that
1) We don't find out directly if the optimum has been set wrong, and
2) We won't measure any progress on problems where the optimum has been set too low.

However, clipping at 0 does make sense for producing plots that are easy to read, since most of the progress often happens in the first trials that constitute the "baseline."

Another issue is with non-monotonic traces. The baseline value should be the best of the first 'n', not the n-th. In the past, all traces were weakly improving, but with inference regret, they need not be.

**This PR:**
* Stops clipping benchmark scores at 100, so they can go over 100.
* Looks at all baseline trials (e.g. the first 5 rather than the 5th) for computing the optimal value
* Changes `compute_score_trace` to take in an `optimal_value` rather than a `problem`, for clarity

Differential Revision: D66976176