Stop clipping benchmark score trace, score non-monotonic traces correctly, etc (#3158)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3158

**Context:**

Clipping benchmark scores at 100 means that
1) We don't find out directly if the optimum has been set wrong, and
2) We won't measure any progress on problems where the optimum has been set too low.

Clipping at 0 is more defensible since it can make plots much easier to read, since most of the progress often happens in the first trials that constitute the "baseline." However, this can be done ex-post; there is no need to lose information at this tage.

Another issue is with non-monotonic traces. The baseline value should be the best of the first 'n', not the n-th. In the past, all traces were weakly improving, but with inference regret, they need not be.

**This PR:**
* Stops clipping benchmark scores to be between 0 and 100.
* Looks at all baseline trials (e.g. the first 5 rather than the 5th) for computing the optimal value
* Changes `compute_score_trace` to take in an `optimal_value` rather than a `problem`, for clarity

Reviewed By: saitcakmak

Differential Revision: D66976176

fbshipit-source-id: 334fea9783aca256728f13e4a821858d28f2a43f