Implements MapKeyToFloat, a subclass of the MetadataToFloat Transform that provides sensible defaults for MapData (#3155)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3155

This adds a specialization of the `MetadataToFloat` Transform, `MapKeyToFloat`, that provides sensible default settings to allow for intercepting map metric data appearing in the ObservationFeatures' metadata.

Additionally, for the purposes of specifying `fixed_features` down the line, when `_transform_observation_feature` is called on an empty `ObservationFeatures` (more specifically, an `ObservationFeatures` with an empty `parameters` dict), it will populate it with the *upper bound* associated with each metadata key in order to fix the candidate suggestion to the highest progression reached so far.

Reviewed By: saitcakmak

Differential Revision: D66945078

fbshipit-source-id: 13fb75adc33a6d86e4e9b68bd10fb650a2215fee