[Docusaurus] Speed up build time using experimental faster option
As part of the v3.6.3 upgrade the Docusaurus team as added opt-in support for a faster build process. This is good for us since since I was experiencing memory-exceeded errors when building the website if there was a large number of versions. Here's a quick benchmark comparison:

> 3.6.2 default settings

[SUCCESS] Generated static files in "build".
[INFO] Use `npm run serve` command to test your build locally.
✨  Done in 61.89s.

> with `@docusaurus/faster` and `"experimental_faster": true`

[SUCCESS] Generated static files in "build".
[INFO] Use `npm run serve` command to test your build locally.
✨  Done in 11.81s.

That's an 80% speed improvement!