Change default MapKeyInfo (#3194)
Summary:
I want to converge around the "step" column name (as we use in the API) as opposed to "steps" -- this may not matter in the long run as we rethink mapdata, but I think now is the time to do it.

cc ltiao who has diffs in flight that may be effected by this.

Lets see if this breaks anything -- shouldnt actually be used anywhere except maybe tests.

Differential Revision: D67413011