Change default MapKeyInfo (#3194)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3194
I want to converge around the "step" column name (as we use in the API) as opposed to "steps" -- this may not matter in the long run as we rethink mapdata, but I think now is the time to do it.

cc ltiao who has diffs in flight that may be effected by this.

Lets see if this breaks anything -- shouldnt actually be used anywhere except maybe tests.

Reviewed By: saitcakmak

Differential Revision: D67413011

fbshipit-source-id: f23cb6c86aed24b643e53d119cb681cbefc5c13e