Speed up Data.filter, part 2 (#3217)
Summary:

`Data.filter` is called `num_metrics x num_trials` times in each `Experiment.fetch_data` call. With large dataframes, A vast majority of time in `Data.filter` is spent in `Data.__init__`, where the bulk of the compute is spent in re-ordering and re-validating and already ordered and validated dataframe. This is wasteful and unnecessary.

This diff adds a `_skip_ordering_and_validation` kwarg to `(Map)Data.__init__` that skips these steps, which is used to immensely speed up `Data.filter`.

{F1974142302}

Reviewed By: bernardbeckerman

Differential Revision: D67720773