adding experiment to modelbridge and using it in TransferToNewSQ (#3220)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3220

Target trial index depends on the current experiment state so it varies and cannot be provided explicitly in the transform configs of TransfrormToNewSQ transform. Thus, we need to compute it inside the transform. Since it depends on the experiment, we needed to add save the experiment on the modelbridge.
- Saved experiment in modelbridge._experiment.
- Using the saved experiment to compute target trial index in TransformToNewSQ.

Reviewed By: saitcakmak

Differential Revision: D67775236

fbshipit-source-id: 65d620e52fbdcf8b52d921d1956bc30e434cee46