Fix trial_as_task for pending points (#3222)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3222

## The Change
If trial indices are not in the level dict, we set TRIAL_PARAM to the max in level_dict, which is the max observed trial index.  We may choose to replace this with the target trial in a follow up.

We then have to ensure the trial_index is not set to something it shouldn't be in `TrialAsTask.untransform_observation_features()` and `ChoiceToNumericChoice.untransform_observation_features()`, which seem to be from generated points and come out as floats.  I'm less certain about why this is.  There may be another error in the modeling stack, but in the meantime it would block candidate generation.

## The Promlem

Reviewed By: lena-kashtelyan

Differential Revision: D67766937

fbshipit-source-id: 145d340a3d5706623cdb0de298aeab3ce24c9b75