Remove mocks from TorchModelBridgeTest.test_best_point
Summary: Context: I'm removing mocks (that do not use `wraps`) from Ax unit tests. There were many in this function. However, I a mock of `gen` in place since this is a ModelBridge test that is checking if the correct arguments get passed to and from `Model`, where the generating happens.

Differential Revision: D68041762