When constructing parameters from lists of names and values, ensure that names and values have the same length
Summary:
Context: There is code like

`parameters={p: float(xbest[i]) for i, p in enumerate(self.parameters)}`

If `xbest` is longer than `self.parameters` (which indicates a problem), the extra elements in `xbest` will be silently discarded. The new behavior will generate an exception.

I did not check everywhere that this might be a problem, just in the places where it was causing trouble for me.

This PR:
* Updates two places where  parameters are constructed to use `dict(zip(..., strict=True))`

Reviewed By: Balandat

Differential Revision: D68035646