Remove mocks from TorchModelBridgeTest.test_best_point (#3228)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3228

Context: I'm removing mocks (that do not use `wraps`) from Ax unit tests. There were many in this function. However, I a mock of `gen` in place since this is a ModelBridge test that is checking if the correct arguments get passed to and from `Model`, where the generating happens.

Reviewed By: saitcakmak

Differential Revision: D68041762

fbshipit-source-id: 44fc70552f01c1ddc2a50798e68abacae24cb9d9