Use double precision and deprecate 'dtype' support in TorchModelBridge (#3234)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3234

**Context:**

Using single-precision data with `TorchModelBridge` may have worked at some point, but it was broken for a long time and is currently broken. Given that it has been broken for so long without complaint, I suspect it was never used. The only test for this mocked out important parts of the modeling layer such that we have never had a test that this runs end-to-end or that the data atually remains in single precision.

I discovered that this doesn't work while attempting to remove mocks from tests.

History:
* `TorchModelBridge` has had an argument `torch_dtype` for as long as Ax has been open-source
* However, this argument was ignored and everything was in double-precision for a long time until D36710539 (PR #977) in May 2022; this was only noticed and fixed because the similar argument `device` was also ignored.
* Scalars extracted from a Numpy array with `np.float32` dtype are `np.float32` and will faily any `checked_cast(float, ...)` or `assert_is_instance(..., float)` checks. I am not sure if this was always the case, but it was true in Numpy 1.26.4.

**Considerations:**

This could probably be fixed by updating a few sites relevant to `TorchModelBridge`, but I don't think it's worthwhile to do that and write the thorough end-to-tend tests that have been lacking given that this feature does not seem to be in demand.

`TorchModel` also has a dtype argument, so it could be worth looking into whether it works and is used.

**This PR:**
* Deprecates `dtype` argument to `TorchModelBridge`
* Makes the modelbridge's dtype always torch.double, as it always was by default
* Removes places where this argument is passed, since passing it is deprecated (and since the default value of double was the only one passed, this wasn't necessary).
* Unrelated: Adds missing plumbing for `fit_abandoned` in `MapTorchModelBridge`

Differential Revision: D68114058

Reviewed By: saitcakmak