Remove mocks from TorchModelBridgeTest.test_TorchModelBridge (#3235)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3235

I'm removing mocks (that do not use `wraps`) from Ax unit tests. There were many in this function. However, I left mocks of `fit` and `gen` in place since this is a ModelBridge test that is checking if the correct arguments get passed to and from `Model`, where the fitting and generating happens. `fit` and `gen` ought to be thoroughly tested in `Model` tests (I haven't checked whether that is the case).

Reviewed By: Balandat

Differential Revision: D68030058

fbshipit-source-id: e349b13c954a0344e77f75633914447e88fb7cb1