Cleanup of type issues in models/model_utils
Summary: A `TypeVar` indicates that multiple objects (e.g. an argument and a return type) are the same type. Using `TypeVar`s with functions that work with either numpy arrays or torch tensors indicates that either everything is a numpy array or everything is a tensor. Otherwise, Pyre has no idea whether arrays might be randomly turning into tensors and vice versa.

Differential Revision: D68332888