typing around ThompsonSampler and its subclasses
Summary:
- Added types in ThompsonSampler and EBAshr.
- Added warnings in ThompsonSampler and EBAshr when the objective is a multi-objective. 
Both ThompsonSampler and EBAshr currently handle multi-objective in the same way by multiplying the samples in the case of TS and shrunken estimates in the case of EBAshr with the objective weights that are +/-1 for metrics in the multi-objective. Given different scaling across metrics, multiplying the metric values with the +/-1 weights does not have a meaningful value.

Reviewed By: saitcakmak

Differential Revision: D68297261