Remove no-longer-used `GenerationNode.fit` kwargs
Summary:
We pass kwargs to ModelSpec.fit in order to pass state extracted when this ModelSpec was used last, back into the ModelSpec, in GenNode.fit (https://fburl.com/code/uic21yuw). 
Along with these, we pass any kwargs manually specified to GenNode.fit, but this behavior is unused in GenStrategy (https://fburl.com/code/20kpa363) and can thus be reaped. That was a remnant of how we used to pass the state from #1, in GenStrategy and not in GenNode (https://www.internalfb.com/diff/D32003890?dst_version_fbid=499298561323086&transaction_fbid=623265442134398) and should be deprecated.


Created from CodeHub with https://fburl.com/edit-in-codehub

Reviewed By: saitcakmak

Differential Revision: D67803329