Don't discard abandoned arms on reduced state (#3261)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3261

Reduced state does still have use for leaving out gen metadata, but I don't know of any use cases with such a huge amount of abandoned arms this will make a difference (considering it can never make more than a 50% difference).

It also affects the contents/functionality of the experiment by making it look like arms are not abandoned and should be used on trials.  And loading and saving with reduced state loses records that arms were abandoned, which creates problems in experimentation.  See N6450800

Reviewed By: lena-kashtelyan

Differential Revision: D68514688

fbshipit-source-id: 9fc1ca8501c0b8dc6d15431bf189e42ebb2f4d39