Add concept of recoverable errors respected by Scheduler (#3262)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3262

The motivation is that some metrics are flaky and we don't want to fail the trial just because we encountered one exception fetching.  Especially trials with `period_of_new_data_after_trial_completion()` > 0.

This alternative to implementing this on the metric is that the set of recoverable errors should be a scheduler option, and it's more a matter of scheduler use case than metric.

Reviewed By: Cesar-Cardoso

Differential Revision: D68273328

fbshipit-source-id: 50d611f1027b97e6e0c761fc976c95819601c1f0