Fix trial index returned by get_best_parameters_from_model_predictions_with_trial_index (#3284)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3284

This method has long been returning the index of the GR that was used to extract the best point, rather than returning the index of the trial that generated the best point. I came across it while looking into something else and figured I'd fix it while at it. This method & other existing best point utilities will likely be replaced in the near future, but it doesn't hurt to it work correctly in the mean time.

Fixes https://github.com/facebook/Ax/issues/1629

Reviewed By: esantorella

Differential Revision: D68834769

fbshipit-source-id: 1f08077496bad98f53521c3f28a8e2b3d2d5609e