[Tutorials] Replace with tutorials from upstream
There are some changes to the tutorials in upstream. Since we moved the files around we can't just resolve merge conflicts. Instead let's just overwrite with the latest, this has the benefit of wiping the data from running the tutorials locally from testing. In the next commit I'll add the code cells again to support google. colab