Make the message clearer to avoid confusion about a scary and lengthy test stack (#3298)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3298

We have a pretty convenient feature in Ax' `TestCase`, to deeply compare unequal objects. This can help a lot if you know how to use it, but folks mentioned to me that because the printout is lengthy, folks sometimes get scared of it and don't read it all the way through. Here's my attempt to make it friendlier : )

For illustration, printout (see how it's actually pretty useful if we look all the way through?):

{F1974874829}

Created from CodeHub with https://fburl.com/edit-in-codehub

Reviewed By: Cesar-Cardoso

Differential Revision: D68987651

fbshipit-source-id: 941b48df7fc900935d4fa564cf1430951c6c9cd6