Updates baseline and optimal values for LCBench early-stopping benchmark problems
Summary:
## Context

For the LCBench *early-stopping* surrogate benchmark problems, we were previously using baseline and optimal values computed specifically for the vanilla (single-output/single-fidelity) LCBench surrogate benchmark problems, which is based on GP surrogate models, while the *early-stopping* surrogate benchmarks are based on random forests. Therefore, we should not expect the baseline and optimal values to transfer.

## Changes

We 1) use N6522262 to compute the baseline values specifically for the  *early-stopping* surrogate benchmarks based on random forests and 2) set the best observed objective value for each dataset as the optimal value. We define `BASELINE_VALUES` and `OPTIMAL_VALUE` dicts consisting of these values and remove the dependency on the analogous dicts defined for the vanilla GP-based LCBench surrogate benchmarks.

Differential Revision: D69150124