Check for feasibility in gen_candidates_scipy and error out for infeasible candidates (#3335)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3335

X-link: https://github.com/pytorch/botorch/pull/2737

As titled. Previously, it was possible to return infeasible candidates to the user, with or without warnings alerting the user to the issue. This diff makes it so that the optimizer will error out when infeasible candidates are generated, so that the user can adjust the setup as needed.

Resolves https://github.com/pytorch/botorch/issues/2708

Also includes a couple lint fixes in optimizer tests.

Reviewed By: esantorella

Differential Revision: D69314159

fbshipit-source-id: ccbc759274b2ae0fc2b38fa19d0e95abf35a2928