Stop infeasible values from contributing non-NaN values to optimization trace (#3340)
Summary:
Pull Request resolved: https://github.com/facebook/Ax/pull/3340

Context: Currently, infeasible trials can increase the optimization trace (T214631737), potentially even beyond what is possible with feasible trials. There is not any generic way to impute a value for them that is worse than any feasible trials, because we generally don't know the worst possible feasible value.

In multi-objective contexts, nothing (weakly) worse than the reference point contributes to hypervolume, so it is fine to preserve the status quo behavior of treating infeasible trials as if they were at the reference point. With a single objective, the best we can do is set the value of that point to -inf or inf.

Since this function takes a cumulative maximum/minimum, it is possible for the trace produced to include non-finite values if the first trial is infeasible.

This function is used in two places:
1. To get a hypervolume trace
2. In benchmarking

(1) is unaffected and I don't forsee any issues with (2), beyond the (intentional) effect that we will start seeing non-finite values. The function that produces baseline values may start producing non-finite values; if those values are used, it could lead to NaNs in scores.

Reviewed By: sdaulton

Differential Revision: D69421757

fbshipit-source-id: e3ca8e161e05492f6ab683005bec370ed7ff87f8