Remove potential SQL deprecation message
Summary:
We added this in D64781681 to get a temperature check on whether or not this feature was being used in OSS. Based on activity on https://github.com/facebook/Ax/issues/2975 the answer is yes.

This is good information for our team to have and will help inform our plans for Ax in the future.

Differential Revision: D69471648